Restructured the robot controller, Doc changes
Moved ROS functionality to a seperate class
Removed repeated checks to environment, it is now implicitly guaranteed
Isolated rosbag and bake read functions, shared between both controllers

Small corrections to litesim and robotsim for new parameter format
Additions to the readme, regarding action in and out
